+++ NEWS.md
@@ -10,6 +10,7 @@
   - Added comprehensive tests for both interactive and non-interactive scenarios
 
 - Removed .dcf file for old Addin (#280)
+- Reworked string handling to use base R and removed `{stringr}`/`{stringi}` dependencies. (#294)
 
 ## Documentation